[add] benchmark config example
This is the input file for all the systems that need to be benchamrked. Every system has different ways to be configured. These can also be given here as input. E.g.: we can add a new key config into every system that need to have configurations other than the default ones.